Rules: New options replace_name to replace %{name} in attributes
If you use the regex feature of Rules/Defaults you may need the real name
of the matched entry in an attribute (for example home of POSIXUser).
You can now enable replace_name for rules or defaults and %{name} will be
replaces in the attribues of the Element before adding them to the target
entry.

This allows you to write something like that in Defaults to assing a default
home directory somewhere else to all users with unset home:

    <POSIXUser name='.*' home='/somewhere/%{name}'/>