SSHbase: Expire metadata cache on pubkey event
SSHbase supply the public keyfiles as additional metadata, so we need to
expire the metadata cache if we get an event for such file. We would only
need to expire the metadata cache for hosts, that reference that exact file
(host or group specific) but we cannot get this information, so that we
simply expire the metadata cache for all hosts.