DBSettings: Automatically upgrade migrations from south to django
On the first use with django-1.7 we manually get the south migration status
and fake apply the django migration to that point. With that users can switch
to django migrations at any point of time.

The code simply takes the name of the last south migration and forwards the
django migrations to that point. Therefor it is required to keep the names
of the migrations in sync.
It is required to keep the migration names of the two systems in sync.