testsuite: Server requires Python2.6
We do not want to run the test for the Server with Python2.4 and Python2.5 and
use the nose-exclude plugin to disable the Server tests based on the directory.
We do not even want to import the modules of the Server, as it might not be
possible anymore (missing dependencies or invalid syntax).

nose-exclude-0.2 breaks compatibility with Python2.5 and below (it is using the
"with open(..):" syntax), so we have to stick to the last version below.